Move the MigrationProgress workflow to its own package.
