+++ src/databricks/labs/ucx/runtime.py
@@ -7,7 +7,7 @@ from pathlib import Path
 from databricks.sdk.config import with_user_agent_extra
 
 from databricks.labs.ucx.__about__ import __version__
-from databricks.labs.ucx.assessment.workflows import Assessment, Failing, MigrationProgress
+from databricks.labs.ucx.assessment.workflows import Assessment, Failing
 from databricks.labs.ucx.contexts.workflow_task import RuntimeContext
 from databricks.labs.ucx.framework.tasks import Task, Workflow, parse_args
 from databricks.labs.ucx.installer.logs import TaskLogger
@@ -18,6 +18,7 @@ from databricks.labs.ucx.hive_metastore.workflows import (
     MigrateHiveSerdeTablesInPlace,
     MigrateExternalTablesCTAS,
 )
+from databricks.labs.ucx.progress.workflows import MigrationProgress
 from databricks.labs.ucx.recon.workflows import MigrationRecon
 from databricks.labs.ucx.source_code.workflows import ExperimentalWorkflowLinter
 from databricks.labs.ucx.workspace_access.workflows import (
